Renaming and relocation of components
