+++ Support.API/appsettings.Development.json
@@ -1,7 +1,4 @@
 {
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost; Port=5432; Database=support; Username=postgres; Password=admin"
-  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
