+++ Support.API/appsettings.Production.json
@@ -1,7 +1,4 @@
 {
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost; Port=5432; Database=support; Username=support; Password=4m!n"
-  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
