+++ Support.API/appsettings.json
@@ -1,4 +1,7 @@
 {
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=SUPPORT_DB_INTERNAL_SERVER; Port=SUPPORT_DB_INTERNAL_PORT; Database=SUPPORT_DB_NAME; Username=SUPPORT_DB_USER; Password=SUPPORT_DB_PASSWORD"
+  },
   "Logging": {
       "LogLevel": {
         "Default": "Information",
