+++ Tests/Support.Bootstrap.Test/Support.API.Services.Test.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Support.Bootstrap\Support.Bootstrap.csproj" />
+    <ProjectReference Include="..\..\Support.API.Services\Support.API.Services.csproj" />
   </ItemGroup>
 
 </Project>
